Extract HIDE_TASKBAR string to TaskbarIntent.ACTION_HIDE_TASKBAR
String com.farmerbb.taskbar.HIDE_TASKBAR is used on many place,
so we should extract it to a constant field to simply the usage.
Because this string is used to Intent, we create a class called
TaskbarIntent, that is similar to system's Intent to store
constant fields used by Intent.

Signed-off-by: utzcoz <utzcoz@outlook.com>