Add basic unit test structure
Selecting robolectric as unit test library to test UI on JVM. We use the
robolectric version 4.3.1, that supports to latest Android Q(final). But
it requires Java 9 to run the tests, so to avoid this problem, we assign
the sdk to 28 in robolectric.properties for almost all tests, and
leaving Android 10 specific features left to ease the usage of
robolectric.

Now we add a test example for U.java getTaskbarPosition function.

For check the tests result, we can execute ./gradlew test on the top
code directory, or use ctrl + shift + F10 in test file to execute unit
test in Android Studio.

Signed-off-by: utzcoz <utzcoz@outlook.com>