Extract method to filter real pinned apps in TaskbarController
Robolectric doesn't support LauncherApps.isPackageEnabled in this shadow
implementation, so we should add a new implementation to support it. And
we should ignore custom shadows from powermock to avoid the problem
that the classloader of custom shadow isn't equals to classloader
used by robolectric. For example, if we don't ignore custom shadow from
powermock, the following casting can't be executed:

TaskbarShadowLauncherApps shadowLauncherApps =
Shadows.shadowOf(launcherApps);

Test: ./gradlew test

Signed-off-by: utzcoz <utzcoz@outlook.com>